Refactor: merged logic to single file
